requirements.txt
asgiref==3.7.2
Django==4.1.13
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
PyJWT==2.8.0
pymongo==3.12.1
pytz==2024.1
sqlparse==0.2.4
typing_extensions==4.10.0
